Externalized usage cache into own package and increased test coverage
